% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circGLM.R
\name{arcDistance}
\alias{arcDistance}
\title{Compute the arc distance between two angular vectors}
\usage{
arcDistance(th1, th2)
}
\arguments{
\item{th1}{The first angular vector in radians.}

\item{th2}{The second angular vector in radians.}
}
\value{
The distance in radians.
}
\description{
Compute the arc distance between two angular vectors
}
