% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cglmShiny.R
\name{cglmShiny}
\alias{cglmShiny}
\title{cglmShiny}
\usage{
cglmShiny()
}
\description{
Run a shiny app interface for this package. Provides a point-and-click interface where the user can load their own data.
}
\examples{

\dontrun{
cglmShiny()
}

}
