% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circGLMClassFunctions.R
\name{circSD}
\alias{circSD}
\title{Comute the Circular Standard Deviation}
\usage{
circSD(x)
}
\arguments{
\item{x}{}
}
\value{
A numeric, the circular standard deviation.
}
\description{
Returns the circular standard deviation of a vector of circular data which is
defined as the square root of minus 2 times the log of the mean resultant
length.
}
