% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circGLMClassFunctions.R
\name{coef.circGLM}
\alias{coef.circGLM}
\title{Extract circGLM Coefficients}
\usage{
\method{coef}{circGLM}(m)
}
\arguments{
\item{m}{A circGLM object.}
}
\value{
A table of coefficients with their corresponding lower and upper bounds.
}
\description{
Create a table of coefficient results from a circGLM object.
}
\examples{
coef(circGLM(rvmc(10, 0, 1)))

}
