% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circGLM.R
\name{fixResultNames}
\alias{fixResultNames}
\title{Fix names for circGLM vector output}
\usage{
fixResultNames(nms)
}
\arguments{
\item{nms}{The original names.}
}
\value{
A character vector of the same length as \code{nms}.
}
\description{
A function to change the names produced in the \code{\link{Rcpp}} code to
more human readable forms.
}
\details{
This is only done if the \code{\link{circGLM}} function is used with
\code{output = "vector"}.
}
