% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circGLMClassFunctions.R
\name{getPMP}
\alias{getPMP}
\title{Obtain posterior model probabilities}
\usage{
getPMP(x, prior_odds = 1)
}
\arguments{
\item{x}{An vector of odds for which to obtain the posterior model probabilities.}

\item{prior_odds}{The prior odds.}
}
\value{
A matrix with two columns, giving the relative probabilities of the first hypothesis versus the second.
}
\description{
Compute posterior model probabilities from odds \code{x} and a prior odds.
}
\examples{
getPMP(3)

}
