% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circGLMClassPrintFunctions.R
\name{print_all.circGLM}
\alias{print_all.circGLM}
\title{Print all results from a circGLM object}
\usage{
print_all.circGLM(m, digits = 3)
}
\arguments{
\item{m}{A circGLM object.}

\item{digits}{Number of digits to display.}
}
\description{
Print all results from a circGLM object
}
\examples{
print(circGLM(rvmc(10, 1, 1)), type = "all")

dat <- generateCircGLMData()
cglmmod <- circGLM(th = dat[, 1], X = dat[, -1])
print(cglmmod, type = "all")
}
\seealso{
\code{\link{"print_text.circGLM"}},
  \code{\link{"print_mcmc.circGLM"}},
  \code{\link{"print_coef.circGLM"}},
  \code{\link{"print.circGLM"}}.
}
