% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circGLMClassPrintFunctions.R
\name{print_coef.circGLM}
\alias{print_coef.circGLM}
\title{Print circGLM coefficients}
\usage{
print_coef.circGLM(m, digits = 3)
}
\arguments{
\item{m}{A \code{circGLM} object.}

\item{digits}{Number of digits to display.}
}
\description{
Print circGLM coefficients
}
\examples{
print(circGLM(rvmc(10, 0, 1)), type = "coef")

dat <- generateCircGLMData()
cglmmod <- circGLM(th = dat[, 1], X = dat[, -1])
print(cglmmod, type = "coef")
}
\seealso{
\code{\link{print_text.circGLM}},
  \code{\link{print_mcmc.circGLM}},
  \code{\link{print_all.circGLM}},
  \code{\link{print.circGLM}}.
}
