% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rvmc}
\alias{rvmc}
\title{Generate a random variate from the von Mises distribution}
\usage{
rvmc(n, mu, kp)
}
\arguments{
\item{n}{The number of random variates required.}

\item{mu}{The required mean direction, mu.}

\item{kp}{The required concentration, kappa.}
}
\value{
A vector of length n containing VM random variates.
}
\description{
This function generates a set of data from the von Mises distribution.
If kappa is very small, return a circular uniform draw, as otherwise the
algorithm will fail.
}
