% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circGLMClassFunctions.R
\name{predict.circGLM}
\alias{predict.circGLM}
\title{Obtain predictions for the circGLM model}
\usage{
\method{predict}{circGLM}(m, newdata)
}
\arguments{
\item{m}{A circGLM object.}

\item{newdata}{A data frame with predictors. The predictors must be the same
as used in the circGLM object and must have the same column names.}
}
\value{
A numeric vector with predictions.
}
\description{
Obtain predictions for the circGLM model
}
\examples{
dat <- generateCircGLMData()
m   <- circGLM(dat[, 1], X = dat[, -1])

# Predictions for the original outcome angles.
predict.circGLM(m)

# Predictions for new data
dat2  <- generateCircGLMData()
predict.circGLM(m, newdata = dat2)
}
