% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circGLMClassFunctions.R
\name{predict_function.circGLM}
\alias{predict_function.circGLM}
\title{Obtain a prediction function from a circGLM object}
\usage{
predict_function.circGLM(m, linkfun = function(x) atanLF(x, 2))
}
\arguments{
\item{m}{A circGLM object.}
}
\value{
A function that takes \code{newdata} as an argument, which must be a
  data frame with predictors. The predictors must be the same as used in the
  circGLM object and must have the same column names.
}
\description{
Obtain a prediction function from a circGLM object
}
\examples{
dat <- generateCircGLMData()
m   <- circGLM(th = dat[, 1], X = dat[, -1])
predfun <- predict_function.circGLM(m)
newd <- generateCircGLMData()

# Predicted values of the new data.
predfun(newd)

}
